Update README.md examples for network, labels
Note that both io.docksal.virtual-host and io.docksal.project-root are required labels for service containers.
Note that io.docksal.project-root is a supported label.
Note that a network needs to be added by the name of "${COMPOSE_PROJECT_NAME}_default".
Provide example docker command and docker-compose yaml in the README.